requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
influxdb-client==1.41.0
multidict==6.0.5
oauthlib==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
rel==0.4.9.8
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
tweepy==4.14.0
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.7.0
yarl==1.9.4
